backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Scanner
semgrep==1.45.0
GitPython==3.1.40

# LLM
openai==1.3.0
google-generativeai==0.3.0

# GitHub
PyGithub==2.1.1

# Database
aiosqlite==0.19.0
asyncpg==0.29.0
redis==5.0.4

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Billing
stripe==7.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
